requirements.txt
streamlit 
pretty_midi 
numpy 
google-generativeai